+++ app/src/main/java/com/wobbz/fartloop/MainActivity.kt
@@ -3,6 +3,7 @@ package com.wobbz.fartloop
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import com.wobbz.fartloop.core.network.NetworkCallbackUtil
 import com.wobbz.fartloop.navigation.FartLooperNavigation
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber