+++ app/src/test/java/com/wobbz/fartloop/NetworkCallbackUtilTest.kt
@@ -9,6 +9,9 @@ import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.LifecycleRegistry
+import com.wobbz.fartloop.core.network.NetworkCallbackUtil
+import com.wobbz.fartloop.core.network.RuleEvaluator
+import com.wobbz.fartloop.core.network.StubRuleEvaluator
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestCoroutineDispatcher
 import kotlinx.coroutines.test.runBlockingTest